requirements.txt
argparse
clip
k_diffusion
sounddevice
torch
wavfile
mtranslate
gtts
openai-whisper
ffmpeg-python
mediapipe
pygame
opencv-python
flask-socketio